requirements.txt
# Schema-first ECS Familiar requirements
copier>=8.0.0
pydantic>=2.0.0
PyYAML>=6.0
jinja2>=3.1.0

# Cold path dependencies (existing)
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
rich>=13.0.0 